Get level layer

Get a geographic layer associated with a level. We can select the geometry and, using boolean parameters, which attributes are included in the layer's table: only the attributes that make up the key and, if applied to a geodimension, inherited attributes to which the prefix of the level where they are defined can be added.

Usage

get_level_layer(gd, level_name, geometry, only_key, inherited, add_prefix)

# S3 method for geodimension
get_level_layer(
  gd,
  level_name = NULL,
  geometry = NULL,
  only_key = FALSE,
  inherited = FALSE,
  add_prefix = TRUE
)

# S3 method for geolevel
get_level_layer(
  gd,
  level_name = NULL,
  geometry = NULL,
  only_key = FALSE,
  inherited = FALSE,
  add_prefix = TRUE
)

Arguments

gd

A geolevel or geodimension object.

level_name

A string.

geometry

A string.

only_key

A boolean.

inherited

A boolean.

add_prefix

A boolean.

Value

A sf object.

See also

Examples

file <- system.file("extdata", "us_layers.gpkg", package = "geodimension")
layer_us_state <- sf::st_read(file, layer = "state", quiet = TRUE)

state <-
  geolevel(name = "state",
           layer = layer_us_state,
           key = "GEOID")

state_ll <- state |>
  get_level_layer("polygon")

county_ll <- gd_us |>
  get_level_layer(level_name = "county",
                  geometry = "polygon",
                  inherited = TRUE)